requirements/base.txt
fastapi==0.100.1
python-multipart==0.0.9
pypdf==4.2.0
aiohttp==3.11.13
httpx==0.28.1
dependency_injector==4.41.0
pydantic_settings==2.0.0
tortoise-orm==0.19.3
aerich==0.6.3
asyncpg==0.28.0
minio==7.2.7